Developing the Triggers and Tweaking the database
There is not much documentation here, but as you can see there are a lot of triggers !
Basically, the whole idea is to keep a record of when dependencies change, and when anything changes.That way the javascript in the client can later say: "okay, I need all this stuff which changed in the past x seconds", and the server can like, say, "this was added and oh look, somebody put in a tag for this thing that did not change", and the client will say "wow! okay I do not have to refresh it, but add in the tag, thanks!"

And they all go have coffee together. Coding can do that

There are over 50 triggers :-(

WIP triggers

WIP triggers

wip triggers

WIP added better change logging