Got in Server Framework, Version insert and update works
Put in a huge system for updating and inserting things. I believe the work will go faster now.
Now, I can just add in simple plugins for each of the other db things. The only complication being when I have to look out for making connections when the children are finished. But the placement is already there. So most of the hard work is done for the api

Smaller things to keep track of

DB change

.184     After Update Triggers now put old page load id into the log instead of new

And I fixed a major bug with empty() in the code in a couple places. 0 was going to null in the db statements