Added js display and container framework (and it tests ok!)
Added a display framework

These are called by containers

A lot of work went into these classes, and they work really well. The container can be owed by a display which is managed by another layer of containers. It allows much to be added quickly for the js gui

A few bug fixes in heartbeat and gokabam

Added a helper function in js typedefs to get the kid prefix given a object type

Added place in test pages to demo the display widgets

Fixed bug in filler where the delted kids array was still showing the internal structure rather than the id string